refactor(engine): extract shared PLATEAU check logic into citygml/common layers
