+++ engine/runtime/action-plateau-processor/src/citygml/mod.rs
@@ -1,0 +1,7 @@
+mod spec;
+mod v2;
+mod v3;
+
+pub(crate) use spec::{CityGmlSpec, GML_DICTIONARY_NS};
+pub(crate) use v2::CITYGML2;
+pub(crate) use v3::CITYGML3;
