+++ engine/runtime/action-plateau-processor/src/common/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod domain_of_definition_validator;
+pub(crate) mod errors;
+mod profile;
+pub(crate) mod udx_folder_extractor;
+pub(crate) use profile::PlateauProfile;
