+++ engine/runtime/action-plateau-processor/src/lib.rs
@@ -1,5 +1,7 @@
 #![recursion_limit = "256"]
 
+pub(crate) mod citygml;
+pub(crate) mod common;
 pub mod mapping;
 pub(crate) mod object_list;
 pub mod plateau3;
