+++ engine/runtime/action-plateau-processor/src/plateau4/mapping.rs
@@ -3,6 +3,7 @@ use std::collections::HashMap;
 use once_cell::sync::Lazy;
 use reearth_flow_runtime::node::{NodeKind, ProcessorFactory};
 
+use super::profile::PLATEAU4;
 use super::{
     attribute_flattener::processor::AttributeFlattenerFactory,
     building_installation_geometry_type_checker::BuildingInstallationGeometryTypeCheckerFactory,
@@ -11,7 +12,6 @@ use super::{
     city_code_extractor::CityCodeExtractorFactory, citygml_mesh_builder::CityGmlMeshBuilderFactory,
     composite_surface_continuity_filter::CompositeSurfaceContinuityFilterFactory,
     destination_mesh_code_extractor::DestinationMeshCodeExtractorFactory,
-    domain_of_definition_validator::DomainOfDefinitionValidatorFactory,
     face_extractor::FaceExtractorFactory,
     flooding_area_surface_generator::FloodingAreaSurfaceGeneratorFactory,
     gml_name_code_space_validator::GmlNameCodeSpaceValidatorFactory,
@@ -20,14 +20,15 @@ use super::{
     object_list_extractor::ObjectListExtractorFactory,
     solid_intersection_test_pair_creator::SolidIntersectionTestPairCreatorFactory,
     tran_xlink_detector::TransportationXlinkDetectorFactory,
-    udx_folder_extractor::UDXFolderExtractorFactory,
     unmatched_xlink_detector::UnmatchedXlinkDetectorFactory,
     unshared_edge_detector::UnsharedEdgeDetectorFactory,
 };
+use crate::common::domain_of_definition_validator::DomainOfDefinitionValidatorFactory;
+use crate::common::udx_folder_extractor::UDXFolderExtractorFactory;
 
 pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Lazy::new(|| {
     let factories: Vec<Box<dyn ProcessorFactory>> = vec![
-        Box::<UDXFolderExtractorFactory>::default(),
+        Box::new(UDXFolderExtractorFactory::new(&PLATEAU4)),
         Box::<MaxLodExtractorFactory>::default(),
         Box::<AttributeFlattenerFactory>::default(),
         Box::<BuildingInstallationGeometryTypeCheckerFactory>::default(),
@@ -40,7 +41,7 @@ pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Laz
         Box::<GmlNameCodeSpaceValidatorFactory>::default(),
         Box::<ObjectListExtractorFactory>::default(),
         Box::<MissingAttributeDetectorFactory>::default(),
-        Box::<DomainOfDefinitionValidatorFactory>::default(),
+        Box::new(DomainOfDefinitionValidatorFactory::new(&PLATEAU4)),
         Box::<UnmatchedXlinkDetectorFactory>::default(),
         Box::<SolidIntersectionTestPairCreatorFactory>::default(),
         Box::<TransportationXlinkDetectorFactory>::default(),
