+++ engine/runtime/action-plateau-processor/src/plateau4/mod.rs
@@ -6,7 +6,6 @@ pub(crate) mod city_code_extractor;
 pub(crate) mod citygml_mesh_builder;
 pub(crate) mod composite_surface_continuity_filter;
 pub(crate) mod destination_mesh_code_extractor;
-pub(crate) mod domain_of_definition_validator;
 pub(crate) mod errors;
 pub(crate) mod face_extractor;
 pub(crate) mod flooding_area_surface_generator;
@@ -15,8 +14,14 @@ pub(crate) mod mapping;
 pub(crate) mod max_lod_extractor;
 pub(crate) mod missing_attribute_detector;
 pub(crate) mod object_list_extractor;
+mod profile;
 pub(crate) mod solid_intersection_test_pair_creator;
 pub(crate) mod tran_xlink_detector;
-pub(crate) mod udx_folder_extractor;
 pub(crate) mod unmatched_xlink_detector;
 pub(crate) mod unshared_edge_detector;
+
+// The PLATEAU4 profile is referenced from plateau4/mapping via `super::profile`.
+// This re-export exists so the common check logic tests can reference it as
+// `crate::plateau4::PLATEAU4`, and is only needed in test builds.
+#[cfg(test)]
+pub(crate) use profile::PLATEAU4;
