+++ engine/runtime/action-plateau-processor/src/plateau6/mapping.rs
@@ -3,13 +3,14 @@ use std::collections::HashMap;
 use once_cell::sync::Lazy;
 use reearth_flow_runtime::node::{NodeKind, ProcessorFactory};
 
-use super::domain_of_definition_validator::DomainOfDefinitionValidatorFactory;
-use super::udx_folder_extractor::UDXFolderExtractorFactory;
+use super::profile::PLATEAU6;
+use crate::common::domain_of_definition_validator::DomainOfDefinitionValidatorFactory;
+use crate::common::udx_folder_extractor::UDXFolderExtractorFactory;
 
 pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Lazy::new(|| {
     let factories: Vec<Box<dyn ProcessorFactory>> = vec![
-        Box::new(UDXFolderExtractorFactory),
-        Box::new(DomainOfDefinitionValidatorFactory),
+        Box::new(UDXFolderExtractorFactory::new(&PLATEAU6)),
+        Box::new(DomainOfDefinitionValidatorFactory::new(&PLATEAU6)),
     ];
     factories
         .into_iter()
