+++ engine/runtime/action-plateau-processor/src/plateau6/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod domain_of_definition_validator;
-pub(crate) mod errors;
-pub(crate) mod mapping;
-pub(crate) mod udx_folder_extractor;
+mod mapping;
+mod profile;
+
+pub(crate) use mapping::ACTION_FACTORY_MAPPINGS;
